requirements.txt
pandas==1.0.1
plotly==4.6.0
dash_core_components==1.9.1
dash==1.11.0
numpy==1.18.1
dash_html_components==1.0.3
pyspark==2.4.5
pymongo==3.10.1
azure-storage-blob==12.3.2
